Move Data Warehouse files to a separate folder. Refactor.
