+++ src/pyprediktorutilities/dwh/dwh_singleton.py
@@ -1,0 +1,6 @@
+from pyprediktorutilities import singleton
+from pyprediktorutilities.dwh import dwh
+
+
+class DwhSingleton(dwh.Dwh, metaclass=singleton.SingletonMeta):
+    pass
