+++ src/pyprediktorutilities/singleton.py
